refactor session 2
half the things aint working yet... the idea with input is that
instead of having a main input function that polls for numerous events
and dispatches to the appropriate systems, it will instead provide
utility functions to know if a particular thing happened, which are
then called by the relevant systems.

for example, when the player is in playing mode, it may continously
poll the input system to chack if the player made a move, instead of
having the input system detect a move and tell the player about
it. the reason being, that if i have systems in the future that need
this info, instead of adding them to the input system, i just have
them call this same function.