dont modify the board from separate threads
to avoid race conditions or other nasty stuff, async processes store
their desired moves in a channel that the controllers poll every
frame, apllying needed moves if present. this way all board operations
are made on the main thread, and no further synchronization is needed
anywhere else in the code. this also meakes it quite easy to play
sounds